requirements.txt
django
django-allauth
cryptography
psycopg2-binary
whitenoise
djangorestframework